requirements.txt
asgiref==3.7.2
coverage==7.2.7
dj-database-url==2.0.0
dj-static==0.0.6
Django==4.2.2
django-adminlte2==0.4.3
django-appconf==1.0.5
django-imagekit==4.1.0
django-stdimage==6.0.1
gunicorn==20.1.0
model-mommy==2.0.0
pilkit==2.0
Pillow==9.5.0
psycopg2-binary==2.9.6
six==1.16.0
sqlparse==0.4.4
static3==0.7.0
typing_extensions==4.7.0
